+++ packages/components/src/components/Modal/Modal.tsx
@@ -1,4 +1,3 @@
-import * as Aria from "react-aria-components";
 import type { PropsWithChildren } from "react";
 import React from "react";
 import styles from "./Modal.module.scss";
@@ -7,11 +6,9 @@ import type { PropsContext } from "@/lib/propsContext";
 import { PropsContextProvider } from "@/lib/propsContext";
 import { TunnelExit, TunnelProvider } from "@mittwald/react-tunnel";
 import type { OverlayController } from "@/lib/controller/overlay";
-import { useOverlayController } from "@/lib/controller/overlay/useOverlayController";
-import { useSyncTriggerState } from "@/components/Modal/hooks/useSyncTriggerState";
-import { OverlayContextProvider } from "@/lib/controller/overlay/context";
 import type { FlowComponentProps } from "@/lib/componentFactory/flowComponent";
 import { flowComponent } from "@/lib/componentFactory/flowComponent";
+import { Overlay } from "@/components/Overlay";
 
 export interface ModalProps
   extends PropsWithChildren,
@@ -20,32 +17,11 @@ export interface ModalProps
   size?: "s" | "m" | "l";
   offCanvas?: boolean;
   controller?: OverlayController;
-  defaultOpen?: boolean;
   slot?: string;
-  /** @internal */
-  reuseControllerFromContext?: boolean;
 }
 
 export const Modal = flowComponent("Modal", (props) => {
-  const {
-    size = "s",
-    offCanvas,
-    controller: stateFromProps,
-    defaultOpen,
-    children,
-    reuseControllerFromContext = false,
-    ...rest
-  } = props;
-
-  const newState = useOverlayController({
-    reuseControllerFromContext,
-    defaultOpen,
-  });
-
-  const state = stateFromProps ?? newState;
-  const isOpen = state.useIsOpen();
-
-  useSyncTriggerState(state);
+  const { size = "s", offCanvas, controller, children, ...rest } = props;
 
   const rootClassName = clsx(
     styles.modal,
@@ -68,26 +44,14 @@ export const Modal = flowComponent("Modal", (props) => {
   };
 
   return (
-    <Aria.ModalOverlay
-      className={styles.overlay}
-      {...rest}
-      isDismissable
-      isOpen={isOpen}
-      onOpenChange={(isOpen) => state.setOpen(isOpen)}
-    >
-      <Aria.Modal className={rootClassName}>
-        <Aria.Dialog className={styles.dialog}>
-          <OverlayContextProvider value={state}>
-            <PropsContextProvider props={propsContext}>
-              <TunnelProvider>
-                <div className={styles.content}>{children}</div>
-                <TunnelExit id="buttons" />
-              </TunnelProvider>
-            </PropsContextProvider>
-          </OverlayContextProvider>
-        </Aria.Dialog>
-      </Aria.Modal>
-    </Aria.ModalOverlay>
+    <Overlay className={rootClassName} controller={controller} {...rest}>
+      <PropsContextProvider props={propsContext}>
+        <TunnelProvider>
+          <div className={styles.content}>{children}</div>
+          <TunnelExit id="buttons" />
+        </TunnelProvider>
+      </PropsContextProvider>
+    </Overlay>
   );
 });