+++ packages/docs/src/content/02-components/status/loading-spinner/examples/default.tsx
@@ -1,3 +1,3 @@
-import { LoadingSpinner } from "@mittwald/flow-react-components/LoadingSpinner";
+import LoadingSpinner from "@mittwald/flow-react-components/LoadingSpinner";
 
 <LoadingSpinner />;