+++ packages/components/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.93](https://github.com/mittwald/flow/compare/0.1.0-alpha.92...0.1.0-alpha.93) (2024-05-02)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.92](https://github.com/mittwald/flow/compare/0.1.0-alpha.91...0.1.0-alpha.92) (2024-05-02)
+
+### Bug Fixes
+
+- **NavigationGroup:** fix hydration error when used in SSR
+  ([4c828c7](https://github.com/mittwald/flow/commit/4c828c7ee95a1eebc070a1ea7c3d3260ff51f6cd))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.91](https://github.com/mittwald/flow/compare/0.1.0-alpha.90...0.1.0-alpha.91) (2024-05-02)
+
+### Features
+
+- **Breadcrumb:** add inverse option to breadcrumb
+  ([a15b46b](https://github.com/mittwald/flow/commit/a15b46b50fb48e3bbe28810f5c27bfddcd5969e8))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.90](https://github.com/mittwald/flow/compare/0.1.0-alpha.89...0.1.0-alpha.90) (2024-05-02)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.89](https://github.com/mittwald/flow/compare/0.1.0-alpha.88...0.1.0-alpha.89) (2024-04-30)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0-alpha.88](https://github.com/mittwald/flow/compare/0.1.0-alpha.87...0.1.0-alpha.88) (2024-04-29)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
 # [0.1.0-alpha.87](https://github.com/mittwald/flow/compare/0.1.0-alpha.86...0.1.0-alpha.87) (2024-04-29)
 
 ### Features